开放一些api

+++ game-utils/src/main/java/com/wjybxx/fastjgame/net/http/HttpPortContext.java
@@ -52,7 +52,7 @@ public class HttpPortContext {
     /**
      * @return session超时时间
      */
-    public int getHttpSessionTimeout() {
+    public long getHttpSessionTimeout() {
         return config.getHttpSessionTimeout();
     }
